+++ NAudioWpfDemo/RealtimeEffectsDemo/RealtimeEffectsView.xaml.cs
@@ -1,0 +1,15 @@
+using System.Windows.Controls;
+
+namespace NAudioWpfDemo.RealtimeEffectsDemo
+{
+    /// <summary>
+    /// Interaction logic for RealtimeEffectsView.xaml
+    /// </summary>
+    public partial class RealtimeEffectsView : UserControl
+    {
+        public RealtimeEffectsView()
+        {
+            InitializeComponent();
+        }
+    }
+}